requirements.txt
google-adk
langchain-community
streamlit